int not float for n

+++ race/src/gap_finder.py
@@ -71,7 +71,7 @@ def callback(data):
 		if params["car_width"] / 2 >= closest_distance:
 			continue
 
-		n = math.ceil(2 * math.asin(params["car_width"] / (4 * closest_distance)) / data.angle_increment)
+		n = int(math.ceil(2 * math.asin(params["car_width"] / (4 * closest_distance)) / data.angle_increment))
 
 		# print(n)
